+++ libhoomd/computes/NeighborListStencil.h
@@ -47,51 +47,80 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-// Maintainer: joaander
+// Maintainer: mphoward
 
-#include "ComputeThermo.h"
+#include "NeighborList.h"
+#include "CellList.h"
+#include "CellListStencil.h"
 
-/*! \file ComputeThermoGPU.h
-    \brief Declares a class for computing thermodynamic quantities on the GPU
+/*! \file NeighborListStencil.h
+    \brief Declares the NeighborListStencil class
 */
 
 #ifdef NVCC
 #error This header cannot be compiled by nvcc
 #endif
 
-#ifndef __COMPUTE_THERMO_GPU_H__
-#define __COMPUTE_THERMO_GPU_H__
+#ifndef __NEIGHBORLISTSTENCIL_H__
+#define __NEIGHBORLISTSTENCIL_H__
 
-//! Computes thermodynamic properties of a group of particles on the GPU
-/*! ComputeThermoGPU is a GPU accelerated implementation of ComputeThermo
+//! Efficient neighbor list build on the CPU with multiple bin stencils
+/*! Implements the O(N) neighbor list build on the CPU using a cell list with multiple bin stencils.
+
+    \sa CellListStencil
     \ingroup computes
 */
-class ComputeThermoGPU : public ComputeThermo
+class NeighborListStencil : public NeighborList
     {
     public:
         //! Constructs the compute
-        ComputeThermoGPU(boost::shared_ptr<SystemDefinition> sysdef,
-                         boost::shared_ptr<ParticleGroup> group,
-                         const std::string& suffix = std::string(""));
-        virtual ~ComputeThermoGPU();
+        NeighborListStencil(boost::shared_ptr<SystemDefinition> sysdef,
+                            Scalar r_cut,
+                            Scalar r_buff,
+                            boost::shared_ptr<CellList> cl = boost::shared_ptr<CellList>(),
+                            boost::shared_ptr<CellListStencil> cls = boost::shared_ptr<CellListStencil>());
+
+        //! Destructor
+        virtual ~NeighborListStencil();
+
+        //! Change the cutoff radius for all pairs
+        virtual void setRCut(Scalar r_cut, Scalar r_buff);
+        
+        //! Set the cutoff radius by pair type
+        virtual void setRCutPair(unsigned int typ1, unsigned int typ2, Scalar r_cut);
+
+        //! Change the underlying cell width
+        void setCellWidth(Scalar cell_width)
+            {
+            m_override_cell_width = true;
+            m_needs_restencil = true;
+            m_cl->setNominalWidth(cell_width);
+            }
+
+        //! Set the maximum diameter to use in computing neighbor lists
+        virtual void setMaximumDiameter(Scalar d_max);
 
     protected:
-        GPUVector<Scalar4> m_scratch;  //!< Scratch space for partial sums
-        GPUVector<Scalar> m_scratch_pressure_tensor; //!< Scratch space for pressure tensor partial sums
-        GPUVector<Scalar> m_scratch_rot; //!< Scratch space for rotational kinetic energy partial sums
-        unsigned int m_block_size;   //!< Block size executed
-        cudaEvent_t m_event;         //!< CUDA event for synchronization
+        //! Builds the neighbor list
+        virtual void buildNlist(unsigned int timestep);
 
-#ifdef ENABLE_MPI
-        //! Reduce properties over MPI
-        virtual void reduceProperties();
-#endif
+    private:
+        boost::shared_ptr<CellList> m_cl;           //!< The cell list
+        boost::shared_ptr<CellListStencil> m_cls;   //!< The cell list stencil
+        bool m_override_cell_width;                 //!< Flag to override the cell width
 
-        //! Does the actual computation
-        virtual void computeProperties();
+        boost::signals2::connection m_rcut_change_conn;     //!< Connection to the cutoff radius changing
+        bool m_needs_restencil;                             //!< Flag for updating the stencil
+        void slotRCutChange()
+            {
+            m_needs_restencil = true;
+            }
+
+        //! Update the stencil radius
+        void updateRStencil();
     };
 
-//! Exports the ComputeThermoGPU class to python
-void export_ComputeThermoGPU();
+//! Exports NeighborListStencil to python
+void export_NeighborListStencil();
 
-#endif
+#endif // __NEIGHBORLISTSTENCIL_H__